# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Foosball tournament REST service.",
        "title": "Go-foosball API",
        "contact": {},
        "version": "0.8"
    },
    "basePath": "/",
    "paths": {
...
        "/games/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Get gamne results",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Game"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
        "/tournaments/{id}/events/game": {
            "get": {
                "produces": [
                    "application/x-json-stream"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Opens a web socket for tournamnent game start events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GameStartEvent"
                        }
                    }
                }
            }
        },
...
        "/tournaments/{id}/games": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Get all games in a tournament",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Game"
                            }
                        }
                    }
                }
            }
        },
        "/tournaments/{id}/games/random": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actions"
                ],
                "summary": "Get random game for a tournament",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Game"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/tournaments/{id}/games/start": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actions"
                ],
                "summary": "Publishes a game start event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
        "/tournaments/{id}/tables/{table}/games": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Submit gamne results",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Table ID",
                        "name": "table",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Submit game results",
                        "name": "game",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/GameResult"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Game"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Color": {
            "type": "object",
            "required": [
                "left",
                "right"
            ],
            "properties": {
                "left": {
                    "type": "string"
                },
                "right": {
                    "type": "string"
                }
            }
        },
...
        "Error": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "Game": {
            "type": "object",
            "required": [
                "created",
                "leftPlayers",
                "leftScore",
                "rightPlayers",
                "rightScore",
                "table",
                "updated"
            ],
            "properties": {
                "created": {
                    "type": "string"
                },
                "leftPlayers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "leftScore": {
                    "type": "integer"
                },
                "rightPlayers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rightScore": {
                    "type": "integer"
                },
                "table": {
                    "$ref": "#/definitions/Table"
                },
                "updated": {
                    "type": "string"
                },
                "winner": {
                    "type": "string"
                }
            }
        },
        "GameResult": {
            "type": "object",
            "required": [
                "leftPlayers",
                "rightPlayers",
                "winner"
            ],
            "properties": {
                "leftPlayers": {
                    "type": "array",
                    "maxItems": 2,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "rightPlayers": {
                    "type": "array",
                    "maxItems": 2,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "winner": {
                    "type": "string",
                    "enum": [
                        "right",
                        "left",
                        "draw"
                    ]
                }
            }
        },
        "GameStartEvent": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "Table": {
            "type": "object",
            "required": [
                "color",
                "created",
                "id",
                "name",
                "updated"
            ],
            "properties": {
                "color": {
                    "$ref": "#/definitions/Color"
                },
                "created": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
...
    }
}